requirements.txt
scikit-learn
pandas
numpy
scipy
nltk
emoji
matplotlib
seaborn
joblib
streamlit
PyYAML
deep-translator
